% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EuclideanSD.R
\docType{data}
\name{nums}
\alias{nums}
\title{nums: the namespace variable.}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
nums
}
\description{
the namespace variable that stores data provided by user.
}
\keyword{datasets}
